pyproject.toml
[tool.poetry]
name = "docker-project"
version = "0.1.0"
description = ""
authors = ["Muzafar-Ali <muzafar.ali.chandio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
